requirements.txt
ariadne==0.26.1 
pymongo==4.11.3 
fastapi==0.115.11 
pydantic==2.10.6 
python-dotenv==1.0.1 
strawberry-graphql==0.262.5 
starlette==0.46.1
motor==3.7.0 
pytest==8.3.5 
uvicorn==0.34.0 
watchgod==0.8.2 